+++ scripts/js-api-tests.js
@@ -3676,7 +3676,7 @@ let transformTests = {
 
   async ts({ esbuild }) {
     const { code } = await esbuild.transform(`enum Foo { FOO }`, { loader: 'ts' })
-    assert.strictEqual(code, `var Foo;\n((Foo2) => {\n  Foo2[Foo2["FOO"] = 0] = "FOO";\n})(Foo || (Foo = {}));\n`)
+    assert.strictEqual(code, `var Foo = /* @__PURE__ */ ((Foo2) => {\n  Foo2[Foo2["FOO"] = 0] = "FOO";\n  return Foo2;\n})(Foo || {});\n`)
   },
 
   async tsx({ esbuild }) {